Forbid narrowing in automap.cpp

+++ common/main/automap.h
@@ -34,7 +34,7 @@ COPYRIGHT 1993-1999 PARALLAX SOFTWARE CORPORATION.  ALL RIGHTS RESERVED.
 #include <array>
 
 namespace dcx {
-extern int Automap_active;
+extern uint8_t Automap_active;
 }
 #ifdef dsx
 namespace dsx {
